% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ind_press.R
\name{check_ind_press}
\alias{check_ind_press}
\title{Check class and variable types of ind and press data}
\usage{
check_ind_press(x, input = "ind")
}
\arguments{
\item{x}{A data object containing only the indicator or pressure
variables (vector, matrix, dataframe or tibble).}

\item{input}{Indication whether x represents the indicator (default) or
pressure variable(s). If x is a vector, 'input' determines the variable
name when coerced to a data frame.}
}
\value{
The function returns the input object as data frame.
}
\description{
\code{check_ind_press} is a helper function for \code{\link{model_trend}}
and \code{\link{ind_init}}. It coerces the data into a data frame and
checks the type of each variable. If any of the variables is not a
number (double or integer) it will return an error message.
}
\seealso{
\code{\link{model_trend}} and the \code{\link{ind_init}}
}
\keyword{internal}
