% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input..R
\name{check_input}
\alias{check_input}
\title{Check class and variable types of input table}
\usage{
check_input(x)
}
\arguments{
\item{x}{A table object (dataframe, tibble, or matrix).}
}
\description{
\code{check_input} is a helper function for \code{\link{model_trend}}
and \code{\link{ind_init}}. It coerce the data table into a data frame
if needed and checks the type of each variable. If any of the variables is not
a number (double or integer) it will return an error message.
}
\seealso{
\code{\link{model_trend}} and the \code{\link{ind_init}}
}
\keyword{internal}
