% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_vec.R
\name{check_input_vec}
\alias{check_input_vec}
\title{Check class and data type of input vector}
\usage{
check_input_vec(x, vec_name)
}
\arguments{
\item{x}{An R object, which should be a numeric vector.}

\item{vec_name}{The name of the input object for the error message.}
}
\value{
The function returns the checked input vector unchanged.
}
\description{
\code{check_input_vec} is a helper function for \code{\link{model_trend}},
\code{\link{ind_init}}, \code{\link{statespace_ed}} and \code{\link{statespace_ch}}.
 It checks whether the time object is an integer vector. If not (a vector
OR integer) it will return an error message.
}
\seealso{
\code{\link{model_trend}}, \code{\link{ind_init}},
 \code{\link{statespace_ed}} and \code{\link{statespace_ch}}
}
\keyword{internal}
