% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ind_ex}
\alias{ind_ex}
\title{Food web indicators of the Central Baltic Sea (Bornholm Basin).}
\format{A data frame with 30 rows and 13 variables. These indicator time series
represent a slight modification of the original time series represented in
Otto et al. (2018)
\describe{
\item{Year}{year given as integer}
\item{TZA}{ln-transformed Total Zooplankton Abundance}
\item{MS}{Mean Size of zooplankton}
\item{rCC}{ln-transformed ratio of Cladoceran to Copepod Abundance}
\item{Cops}{ln-transformed Copepod biomass}
\item{Micro}{ln-transformed Biomass of Microphageous zooplankton}
\item{rZPPP}{ratio of Zooplankton to Phytoplankton biomass}
\item{Sprat}{ln-transformed Sprat abundance}
\item{Herring}{ln-transformed Herring abundance}
\item{Stickle}{ln-transformed Stickleback CPUE in numbers}
\item{Cod}{ln-transformed Cod CPUE in numbers}
\item{SPF}{ln-transformed Small Predatory Fish (i.e. sprat and herring <10cm)}
\item{LPF}{ln-transformed Large Predatory Fish (i.e. Cod >38cm)}
}}
\usage{
ind_ex
}
\description{
Food web indicators of the Central Baltic Sea (Bornholm Basin).
}
\keyword{datasets}
