% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ind_init_ex}
\alias{ind_init_ex}
\title{Output tibble from the \code{\link{ind_init}} function}
\format{
A data frame with 84 rows and 10 variables:
\describe{
  \item{\code{id}}{Numerical IDs for the IND~press combinations.}
  \item{\code{ind}}{Indicator names.}
  \item{\code{press}}{Pressure names.}
  \item{\code{ind_train}}{A list-column with indicator values of the training data.}
  \item{\code{press_train}}{A list-column with pressure values of the training data.}
  \item{\code{time_train}}{train data from year}
  \item{\code{ind_test}}{A list-column with indicator values of the test data.}
  \item{\code{press_test}}{A list-column with pressure values of the test data.}
  \item{\code{time_test}}{test data from year}
  \item{\code{train_na}}{logical; indicates the joint missing values in the training
  IND and pressure data. That includes the original NAs as well as randomly selected
  test observations that are within the training period}
}
}
\usage{
ind_init_ex
}
\description{
This is an example output tibble from the \code{ind_init} function applied
on the Central Baltic Sea food web indicator demonstration data.
}
\keyword{datasets}
