% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv_thresh_gamm.R
\name{loocv_thresh_gamm}
\alias{loocv_thresh_gamm}
\title{Leave-One-Out Cross-Validation (LOOCV) procedure for (threshold-)GAMMs
- NOT WORKING YET!}
\usage{
loocv_thresh_gamm(model, ind_vec, press_vec, t_var, name_t_var, k, a, b, time)
}
\arguments{
\item{model}{A single GAM object from the model tibble needed to extract
the family and the link function.}

\item{ind_vec}{A vector with the IND training observations (including or excluding
defined outliers).}

\item{press_vec}{A vector with the training observations (including or excluding
defined outliers) of pressure 1 (i.e. the orginal significant pressure in the
GAM(M)).}

\item{t_var}{A vector with the training observations (including or excluding
defined outliers) of the threshold variable (i.e. a second pressure variable).}

\item{name_t_var}{The name of the threshold variable (pressure 2). t_var will be
named after this string in the model formula.}

\item{k}{Choice of knots (for the smoothing function \code{\link{s}}); the
default is 4 to avoid over-parameterization.}

\item{a}{The lower quantile value of the selected threshold variable, which
the estimated threshold is not allowed to exceed; the default is 0.2.}

\item{b}{The upper quantile value of the selected threshold variable, which
the estimated threshold is not allowed to exceed; the default is 0.8.}

\item{time}{A vector containing the actual time series.}
}
\description{
\code{loocv_thresh_gamm} applies a LOOCV on a threshold-GAMM and its
corresponding GAMM and returns TRUE if the threshold-GAMM has a lower
estimate, else FALSE (see for more infos on the LOOCV procedure
the details section in \code{\link{test_interaction}}).
}
\examples{
#rm(list = ls())
\dontrun{
ind_vec <- ind_init_ex$ind_train[[67]]
press_vec <- ind_init_ex$press_train[[67]]
t_var <- ind_init_ex$press_train[[63]]
name_t_var <- "Fher"
time <- ind_init_ex$time_train[[67]]
model <- merge_models_ex$model[[67]]
k <- 4; a <- 0.2; b <- 0.8

#takes some time!
dat <- loocv_thresh_gamm(model, ind_vec, press_vec, t_var, name_t_var, k, a, b, time)
}

}
\seealso{
\code{\link{thresh_gamm}} which creates a threshold-GAMM object and
\code{\link{test_interaction}} which applies thresh_gamm and loocv_thresh_gamm
}
