% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_trend_ex}
\alias{model_trend_ex}
\title{Model output tibble from the \code{\link{model_trend}} function}
\format{A data frame with 12 rows and 9 variables:
\describe{
  \item{\code{ind_id}}{Indicator IDs.}
  \item{\code{ind}}{Indicator names.}
  \item{\code{p_val}}{The p values for the smoothing term (here time).}
  \item{\code{model}}{A list-column of indicator-specific gam objects.}
  \item{\code{ind_train}}{A list-column with indicator values of the training data.}
  \item{\code{time_train}}{A list-column with the time values (e.g. years) of the
             training data.}
  \item{\code{pred}}{A list-column with indicator values predicted from the GAM
             for the training period.}
  \item{\code{ci_up}}{A list-column with the upper 95\% confidence interval of
             predicted indicator values.}
  \item{\code{ci_low}}{A list-column with the lower 95\% confidence interval of
             predicted indicator values.}
}}
\usage{
model_trend_ex
}
\description{
This is an example output tibble from the \code{model_trend} function applied on the
Central Baltic Sea food web indicator demonstration data.
}
\keyword{datasets}
