% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\description{
INDperform makes heavy use of the tidy data principles.
Therefore it is advisable to import the pipe operator \code{\%>\%}
from magrittr.
}
\keyword{internal}
