% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_text.R
\name{place_text}
\alias{place_text}
\title{Place annotations to a specified position}
\usage{
place_text(x, y, x_prop = 0.1, y_prop = 0.1, pos)
}
\arguments{
\item{x}{The complete vector of x values in the ggplot.}

\item{y}{The complete vector of y values in the ggplot.}

\item{pos}{Specifies the corner in which to place the annotation.
This must be one of "topleft", "topright", "bottomleft" and
"bottomright"; default is "topleft".}

\item{x_range}{The required width of the text box. Numeric value
between 0 and 1; default is 0.1.}

\item{y_range}{The required height of the text box. Numeric value
between 0 and 1; default is 0.2.}
}
\value{
The function returns a vector with the x and y coordinates
depending on the selected position.
}
\description{
\code{place_text} is a helper function for \code{\link{plot_trend}}
and \code{\link{plot_model}} to generate the x and y coordinates
for placing annotations in the ggplot depending on the specified
position.
}
\seealso{
\code{\link{plot_trend}}, \code{\link{plot_model}}
}
\keyword{internal}
