% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plot_diagnostics.R
\name{plot_outline}
\alias{plot_outline}
\alias{plot_cook}
\alias{plot_acf}
\alias{plot_pacf}
\alias{plot_resid}
\alias{plot_qq}
\alias{plot_gcvv}
\alias{plot_all_diag}
\alias{plot_empty}
\title{Helper functions for plot_model}
\usage{
plot_outline()

plot_cook(values, title)

plot_acf(x_var, y_var, title)

plot_pacf(x_var, y_var, title)

plot_resid(model_fitted, model_resid, title)

plot_qq(model_resid, theo_quan, title)

plot_gcvv(x_var, y_var, lab, best_t_val, title)

plot_all_diag(p1, p2, p3, p4, p5, p6, title)

plot_empty()
}
\description{
These are the actual plotting functions for the 6 plots shown in the wrapper
function \code{\link{plot_diagnostics}}.
}
\seealso{
\code{\link{plot_diagnostics}}
}
\keyword{internal}
