% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plot_spiechart.R
\name{plot_spie}
\alias{plot_spie}
\title{Helper function for plot spiechart}
\usage{
plot_spie(split_input, scale, parting, cat, summary_tbl, ground, n_c8_c11,
  n_c9_c10, col_crit8_11, x_ring1, theme_infog, edge = "grey30", ind,
  lab_size, title_size)
}
\description{
This is the actual plotting function used by the wrapper
function \code{\link{plot_spiechart}} for each indicator.
}
\seealso{
\code{\link{plot_spiechart}}
}
\keyword{internal}
