% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{press_ex}
\alias{press_ex}
\title{Environmental variables representing pressures for pelagic food webs in
the Central Baltic Sea (Bornholm Basin).}
\format{
A data frame with 30 rows and 8 variables. These indicator time series
represent a slight modification of the original time series represented in
Otto et al. (2018)
\describe{
\item{Year}{year given as integer}
\item{Tsum}{temperature summer (in degree C)}
\item{Swin}{salinity winter}
\item{Pwin}{phosphate in winter (in mg m^-3^)}
\item{Nwin}{nitrogen winter (in mg m^-3^)}
\item{Fsprat}{fishing mortality of sprat}
\item{Fher}{fishing mortality of herring}
\item{Fcod}{fishing mortality of cod}
}
}
\usage{
press_ex
}
\description{
Environmental variables representing pressures for pelagic food webs in
the Central Baltic Sea (Bornholm Basin).
}
\keyword{datasets}
