% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{press_type_ex}
\alias{press_type_ex}
\title{Pressure variables and their associated pressure types from our example data}
\format{A data frame with 7 rows and 2 variables:
\describe{
\item{press}{pressure name}
\item{press_type}{corresponding pressure type}
}}
\usage{
press_type_ex
}
\description{
Pressure variables and their associated pressure types from our example data
}
\keyword{datasets}
