% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_output_tbl.R
\name{sort_output_tbl}
\alias{sort_output_tbl}
\title{Rearrange columns in output tibble from IND~pressure modelling functions}
\usage{
sort_output_tbl(mod_tbl)
}
\arguments{
\item{mod_tbl}{A tibble, e.g. tibbles generated by the IND~pressure
modelling functions}
}
\value{
The function returns a re-arranged input tibble.
}
\description{
\code{sort_output_tbl} is a helper function for all IND~pressure
modelling functions and rearranges columns so that all output tibbles show
the same order of variables. This is a generic function in the sense that
all other variables not  considered in the sorting list here are placed at
the end. That way, any modified tibble can be used as input.
}
\keyword{internal}
