% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_gamm.R
\name{thresh_gamm}
\alias{thresh_gamm}
\title{Fits a GAMM with a threshold-formulation - NOT WORKING YET!}
\usage{
thresh_gamm(model, ind_vec, press_vec, t_var, name_t_var, k, a, b, time)
}
\arguments{
\item{model}{A single GAM object from the model tibble needed to extract
the family and the link function.}

\item{ind_vec}{A vector with the IND training observations (including or excluding
defined outliers).}

\item{press_vec}{A vector with the training observations (including or excluding
defined outliers) of pressure 1 (i.e. the orginal significant pressure in the
GAM(M)).}

\item{t_var}{A vector with the training observations (including or excluding
defined outliers) of the threshold variable (i.e. a second pressure variable).}

\item{name_t_var}{The name of the threshold variable (pressure 2). t_var will be
named after this string in the model formula.}

\item{k}{Choice of knots (for the smoothing function \code{\link{s}}); the
default is 4 to avoid over-parameterization.}

\item{a}{The lower quantile value of the selected threshold variable, which
the estimated threshold is not allowed to exceed; the default is 0.2.}

\item{b}{The upper quantile value of the selected threshold variable, which
the estimated threshold is not allowed to exceed; the default is 0.8.}

\item{time}{A vector containing the actual time series.}
}
\value{
The function returns a \code{gamm} object with the additional class "\code{tgamm}".
The object has four additional elements in the \code{gam} list :
\describe{
  \item{\code{mr}}{The threshold value of the best threshold-GAM.}
  \item{\code{mgcv}}{The GCV of the best threshold-GAM.}
  \item{\code{gcvv}}{A vector of the GCV values of all fitted threshold-GAMs.}
  \item{\code{t_val}}{A vector of all tested threshold values within the
             boundaries set by the lower and upper quantiles.}
  \item{\code{train_na}}{A logical vector indicating missing values.}
}
}
\description{
\code{thresh_gamm} fits a Generalized Additive Mixed Model (GAMM) with a threshold
formulation. The threshold value is estimated from the data and chosen by minimizing
the GCV score (termed 'gcvv' in the threshold-GAMM object) over an interval defined
by the lower and upper quantiles (see the \code{a} and \code{b} arguments
respectively) of the threshold variable. The correlation structure included is based
on the original IND~pressure GAMM.
}
\examples{
#rm(list = ls())
\dontrun{
ind_vec <- ind_init_ex$ind_train[[43]]
press_vec <- ind_init_ex$press_train[[43]]
t_var <- ind_init_ex$press_train[[2]]
name_t_var <- "Ssum"
time <- ind_init_ex$time_train[[43]]
model <- model_gamm_ex$model[[2]]
k <- 4; a <- 0.2; b <- 0.8
dat <- thresh_gamm(model, ind_vec, press_vec, t_var, name_t_var, k, a, b, time)
}
}
\seealso{
\code{\link{test_interaction}} which applies thresh_gamm
 and \code{\link{loocv_thresh_gamm}}
}
