% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{press_ex}
\alias{press_ex}
\title{Environmental variables representing pressures for pelagic food webs in
the Central Baltic Sea (Bornholm Basin).}
\format{A data frame with 30 rows and 12 variables. These indicator time series
represent a slight modification of the original time series represented in
Otto et al. (2018)
\describe{
\item{Year}{year given as integer}
\item{Tsum}{temperature summer}
\item{Ssum}{salinity summer}
\item{Swin}{salinity winter}
\item{Psum}{phosphate in summer}
\item{Pwin}{phosphate in winter}
\item{Nsum}{nitrogen in summer}
\item{Nwin}{nitrogen winter}
\item{rNPsum}{ratio of nitrogen to phasphate in summer}
\item{Fsprat}{fishing mortality of sprat}
\item{Fher}{fishing mortality of herring}
\item{Fcod}{fishing mortality of cod}
}}
\source{
see dummy.R for details
}
\usage{
press_ex
}
\description{
Environmental variables representing pressures for pelagic food webs in
the Central Baltic Sea (Bornholm Basin).
}
\keyword{datasets}
